@@ -16,7 +16,7 @@
 
 [tool.poetry]
 name = "splunk_add_on_ucc_framework"
-version = "5.69.0"
+version = "5.69.1"
 description = "Splunk Add-on SDK formerly UCC is a build and code generation framework"
 license = "Apache-2.0"
 authors = ["Splunk <addonfactory@splunk.com>"]
@@ -33,8 +33,6 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Python Modules",
     "Topic :: Software Development :: Code Generators",
     "License :: OSI Approved :: Apache Software License",
-    "Programming Language :: Python :: 3.7",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -46,15 +44,15 @@ classifiers = [
 "Bug Tracker" = "https://github.com/splunk/addonfactory-ucc-generator/issues"
 
 [tool.poetry.dependencies]
-python = ">=3.7,<3.14"
-jinja2 = ">=2,<4"
+python = ">=3.9,<3.14"
+jinja2 = ">=3.1.6,<4"
 addonfactory-splunk-conf-parser-lib = "^0.4.3"
 dunamai = "^1.22.0"
-jsonschema = "^4.4.0"
+jsonschema = "^4.25.0"
 PyYAML = "^6.0"
 defusedxml = "^0.7.1"
 colorama = "^0.4.6"
-packaging = ">=23.0"
+packaging = ">=25.0"
 splunk-appinspect = { version = "^3.10.3", extras = ["validate"], optional = true }
 certifi = "^2025.6.15"
 
@@ -63,18 +61,17 @@ validate = ["splunk-appinspect"]
 
 
 [tool.poetry.group.dev.dependencies]
-mkdocs = "^1.4.2"
-importlib-metadata = {version="*", python="<3.8"}
+mkdocs = "^1.6.1"
 pytest = "^7.2.1"
 pytest-splunk-addon = "^6.0.0"
-pytest-splunk-addon-ui-smartx = "^5.3.2"
+pytest-splunk-addon-ui-smartx = "^5.3.4"
 pytest-rerunfailures = "^11.1.1"
-mkdocs-material = "^9.1.3"
+mkdocs-material = "^9.6.16"
 mkdocstrings = {version=">=0", extras=["python"]}
 mkdocs-print-site-plugin = "^2.3.6"
-pytest-cov = "^4.0.0"
+pytest-cov = "^6.0.0"
 covdefaults = "^2.3.0"
-xmldiff = "^2.6.3"
+xmldiff = "^2.7.0"
 pytest-split-tests = "^1.1.0"
 
 [tool.poetry.scripts]
