@@ -25,8 +25,6 @@ export { CustomControlBase } from './components/CustomControl/CustomControlBase'
 
 export { CustomTabBase } from './components/CustomTab/CustomTabBase';
 
-export { CustomMenuBase } from './components/CustomMenu/CustomMenuBase';
-
 export { CustomCellBase } from './components/table/CustomTableCellBase';
 
 export { CustomRowBase } from './components/table/CustomRowBase';
@@ -36,3 +34,8 @@ export { uccInit } from './pages/UccInit';
 export type { Tab } from './components/CustomTab/CustomTab.types';
 
 export type { RowDataFields } from './context/TableContext';
+
+export {
+    renderConfigurationPage,
+    renderInputsPage,
+} from './UccTestCommand/uiUccTests/testUccPages';
